BOL-375 refactor: GroupSearchUiModel data package로 이동
