+++ presentation/src/main/java/com/yapp/bol/presentation/view/group/search/GroupSearchViewModel.kt
@@ -7,7 +7,7 @@ import androidx.paging.cachedIn
 import androidx.paging.insertSeparators
 import androidx.paging.map
 import com.yapp.bol.domain.usecase.group.SearchGroupByKeywordUseCase
-import com.yapp.bol.presentation.data.GroupSearchUiModel
+import com.yapp.bol.presentation.model.GroupSearchUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
